+++ web_app/package.json
@@ -9,6 +9,7 @@
     "license": "ISC",
     "dependencies": {
         "body-parser": "^1.20.1",
+        "cors": "^2.8.5",
         "express": "^4.18.2",
         "moment": "^2.29.4",
         "mysql": "^2.18.1",
